+++ lib/validates_lengths_from_database/railtie.rb
@@ -1,7 +1,7 @@
 module ValidatesLengthsFromDatabase
   if defined?(Rails::Railtie)
     require "rails"
-    
+
     class Railtie < Rails::Railtie
       initializer "validates_lengths_from_database.extend_active_record" do
         ActiveSupport.on_load(:active_record) do
@@ -10,7 +10,7 @@ module ValidatesLengthsFromDatabase
       end
     end
   end
-  
+
   class Railtie
     def self.insert
       ActiveRecord::Base.send(:include, ValidatesLengthsFromDatabase)
